src/requirements.txt
requests
python-decouple
openai